refactor(ipc): extract macOS-permission commands into commands/macos.rs (#82)
Third extraction under #82. After meeting (PR #168) and models
(PR #179), the macOS-permission surface is the next natural
seam: three commands (open_macos_privacy_pane, diagnose_macos_permissions,
reset_macos_permissions), two result types
(MacosPermissionDiagnostic, MacosPermissionResetResult), the
MACOS_BUNDLE_ID const, and they're already cfg-gated by platform
so the move is self-contained.

Move:

- New `commands/macos.rs` with the macOS block in full.
- Module declaration `pub mod macos;` in `commands/mod.rs`.
- `lib.rs` `generate_handler!` invocations updated from
  `ipc::commands::open_macos_privacy_pane` (etc.) to
  `ipc::commands::macos::*`. Macro path-sensitivity per
  learnings.md 2026-04-25.

`commands/mod.rs` drops from 1586 → 1341 LOC. With meeting +
models + macos all extracted, the remaining domains are smaller
(history / replacements / vocabulary at ~50 LOC each, first-run /
ptt-config) and not worth standalone files — they read fine
inline with the dictation hot path.

### Smoke

- `cargo build`, `cargo clippy --all-targets -- -D warnings` clean
- `cargo test --lib` 214 passed
- `npm run tauri dev` launches; macOS permissions diagnostic
  still resolves on the Settings → Permissions tab.
