+++ src-tauri/src/lib.rs
@@ -199,9 +199,9 @@ pub fn run() {
             ipc::commands::reset_first_run,
             ipc::commands::ptt_get_config,
             ipc::commands::ptt_set_config,
-            ipc::commands::open_macos_privacy_pane,
-            ipc::commands::diagnose_macos_permissions,
-            ipc::commands::reset_macos_permissions,
+            ipc::commands::macos::open_macos_privacy_pane,
+            ipc::commands::macos::diagnose_macos_permissions,
+            ipc::commands::macos::reset_macos_permissions,
             ipc::commands::meeting::meeting_sessions_list,
             ipc::commands::meeting::meeting_session_get,
             ipc::commands::meeting::meeting_session_delete,
